[Application] Move Config class to Data namespace.
