+++ src/Valkyrja/Application/Data/Config.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Application;
+namespace Valkyrja\Application\Data;
 
 use Valkyrja\Container\Contract\Service;
 use Valkyrja\Container\Provider\Provider;
