+++ src/Valkyrja/Application/Entry/App.php
@@ -13,9 +13,9 @@ declare(strict_types=1);
 
 namespace Valkyrja\Application\Entry;
 
-use Valkyrja\Application\Config;
 use Valkyrja\Application\Contract\Application;
 use Valkyrja\Application\Data;
+use Valkyrja\Application\Data\Config;
 use Valkyrja\Application\Env;
 use Valkyrja\Application\Exception\RuntimeException;
 use Valkyrja\Application\Valkyrja;
