+++ tests/Unit/Application/ApplicationTest.php
@@ -17,9 +17,9 @@ use GuzzleHttp\Client as Guzzle;
 use Valkyrja\Application\Cli\Command\CacheCommand;
 use Valkyrja\Application\Cli\Command\ClearCacheCommand;
 use Valkyrja\Application\Component;
-use Valkyrja\Application\Config;
 use Valkyrja\Application\Contract\Application;
 use Valkyrja\Application\Data;
+use Valkyrja\Application\Data\Config;
 use Valkyrja\Application\Env;
 use Valkyrja\Application\Exception\RuntimeException;
 use Valkyrja\Application\Valkyrja;
